Move uri_properties to Ansible module
